requirements.txt
selenium>=4.15.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv==1.0.1
rich==13.7.1
